[crispy] Move threadName() to crispy library
